Randomizes enemy spawn position, velocity value and angle, implements edge bouncing

+++ assignment-2/src/Config.h
@@ -43,7 +43,7 @@ public:
 	{
 		unsigned int radius;
 		unsigned int collisionRadius;
-		Interval<float> velocity;
+		Interval<int> velocity;
 		sf::Color outlineColor;
 		unsigned int outlineThickness;
 		Interval<int> vertices;
